+++ .next/static/chunks/webpack.js
@@ -192,7 +192,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	!function() {
-/******/ 		__webpack_require__.h = function() { return "640c04fe009387f2"; }
+/******/ 		__webpack_require__.h = function() { return "b68990829b21a025"; }
 /******/ 	}();
 /******/ 	
 /******/ 	/* webpack/runtime/global */